Move knowpro/storage.py to storage/memorystorage.py
