+++ python/ta/test/test_secindex_storage_integration.py
@@ -3,7 +3,7 @@ import pytest
 
 from fixtures import needs_auth, storage, embedding_model  # type: ignore  # It's used!
 from typeagent.knowpro.secindex import ConversationSecondaryIndexes
-from typeagent.knowpro.storage import MemoryStorageProvider
+from typeagent.storage.memorystorage import MemoryStorageProvider
 
 
 @pytest.mark.asyncio
